+++ alacritty_terminal/src/tty/unix.rs
@@ -21,7 +21,8 @@ use mio::unix::EventedFd;
 use nix::pty::openpty;
 #[cfg(any(target_os = "linux", target_os = "macos"))]
 use nix::sys::termios::{self, InputFlags, SetArg};
-use signal_hook::{self as sighook, iterator::Signals};
+use signal_hook::consts::signal::*;
+use signal_hook_mio::v0_6::Signals;
 
 use crate::config::{Config, Program};
 use crate::event::OnResize;
@@ -211,12 +212,12 @@ pub fn new<C>(config: &Config<C>, size: &SizeInfo, window_id: Option<usize>) -> 
             libc::close(slave);
             libc::close(master);
 
-            libc::signal(libc::SIGCHLD, libc::SIG_DFL);
-            libc::signal(libc::SIGHUP, libc::SIG_DFL);
-            libc::signal(libc::SIGINT, libc::SIG_DFL);
-            libc::signal(libc::SIGQUIT, libc::SIG_DFL);
-            libc::signal(libc::SIGTERM, libc::SIG_DFL);
-            libc::signal(libc::SIGALRM, libc::SIG_DFL);
+            libc::signal(SIGCHLD, libc::SIG_DFL);
+            libc::signal(SIGHUP, libc::SIG_DFL);
+            libc::signal(SIGINT, libc::SIG_DFL);
+            libc::signal(SIGQUIT, libc::SIG_DFL);
+            libc::signal(SIGTERM, libc::SIG_DFL);
+            libc::signal(SIGALRM, libc::SIG_DFL);
 
             Ok(())
         });
@@ -228,7 +229,7 @@ pub fn new<C>(config: &Config<C>, size: &SizeInfo, window_id: Option<usize>) -> 
     }
 
     // Prepare signal handling before spawning child.
-    let signals = Signals::new(&[sighook::SIGCHLD]).expect("error preparing signal handling");
+    let signals = Signals::new(&[SIGCHLD]).expect("error preparing signal handling");
 
     match builder.spawn() {
         Ok(child) => {
@@ -328,7 +329,7 @@ impl EventedPty for Pty {
     #[inline]
     fn next_child_event(&mut self) -> Option<ChildEvent> {
         self.signals.pending().next().and_then(|signal| {
-            if signal != sighook::SIGCHLD {
+            if signal != SIGCHLD {
                 return None;
             }